Use consistent navigation view, titles, and dismiss logic in POSCouponCreationSheet
