+++ WooCommerce/Classes/POS/Presentation/Coupons/POSCouponCreationSheet.swift
@@ -1,0 +1,77 @@
+import SwiftUI
+
+extension View {
+    func posCouponCreationSheet(
+        isPresented: Binding<Bool>,
+        onSuccess: @escaping () -> Void
+    ) -> some View {
+        modifier(POSCouponCreationSheetModifier(isPresented: isPresented, onSuccess: onSuccess))
+    }
+}
+
+private struct POSCouponCreationSheetModifier: ViewModifier {
+    @Binding var isPresented: Bool
+    let onSuccess: () -> Void
+
+    @State private var selectedType: POSCouponDiscountType?
+
+    func body(content: Content) -> some View {
+        content
+            .sheet(item: $selectedType) { (posDiscountType: POSCouponDiscountType) in
+                var view = AddEditCoupon(.init(discountType: posDiscountType.discountType, onSuccess: { _ in
+                    onSuccess()
+                }))
+
+                view.dismissHandler = {
+                    selectedType = nil
+                }
+
+                view.onDisappear = {
+                    selectedType = nil
+                }
+
+                return view
+                    .interactiveDismissDisabled()
+            }
+            .sheet(isPresented: $isPresented) {
+                let command = DiscountTypeBottomSheetListSelectorCommand(selected: nil) { type in
+                    selectedType = POSCouponDiscountType(discountType: type)
+                }
+
+                NavigationView {
+                    BottomSheetListSelector(
+                        viewProperties: BottomSheetListSelectorViewProperties(),
+                        command: command,
+                        onDismiss: { _ in
+                            isPresented = false
+                        }
+                    )
+                    .navigationBarTitleDisplayMode(.large)
+                    .navigationTitle(Localization.selectCouponTypeTitle)
+                    .toolbar {
+                        ToolbarItem(placement: .cancellationAction) {
+                            Button(Localization.selectCouponCancelButtonTitle) {
+                                isPresented = false
+                            }
+                        }
+                    }
+                }
+                .navigationViewStyle(.stack)
+                .interactiveDismissDisabled()
+            }
+    }
+}
+
+private enum Localization {
+    static let selectCouponTypeTitle = NSLocalizedString(
+        "pos.couponCreationSheet.selectCoupon.title",
+        value: "Create coupon",
+        comment: "A title for the view that selects the type of coupon to create"
+    )
+
+    static let selectCouponCancelButtonTitle = NSLocalizedString(
+        "pos.couponCreationSheet.selectCoupon.cancel",
+        value: "Cancel",
+        comment: "A button that dismisses coupon creation sheet"
+    )
+}
